requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
et-xmlfile==1.0.1
httplib2==0.18.1
jdcal==1.4.1
numpy==1.19.5
openpyxl==3.0.5
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
soupsieve==2.1
xlrd==1.2.0
